Updating practice materials

+++ fibonacci/main.go
@@ -3,10 +3,10 @@ package main
 import "fmt"
 
 func main() {
-	var a int64 = 1
-	var b int64 = 1
-	var c int64 = 0
-	var sum int64 = 0
+	var a int = 1
+	var b int = 1
+	var c int = 0
+	var sum int = 0
 	for ; sum <= 4000000; {
 		c = a + b
 		if c % 2 == 0 {
